+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Account for `NULL` value in left hand side of formula in `specify()` (#156) and `type` in `generate()` (#157).
 - Update documentation code to follow tidyverse style guide (#159).
+- Remove help page for internal `set_params()` (#165).
+- Fully use {tibble} (#166).
 
 # infer 0.3.1